Updated App.js to use temporary routes and added placeholder components
