+++ src/App.js
@@ -1,99 +1,25 @@
 import React from 'react';
 import SideBar from './SideBar'
-import {BarChart, ScatterPlot, LineChart} from 'react-d3-components'
-import data from './assets/MOCK_DATA_150.json'
+import Charts from './Charts'
+import Blank from './Blank'
+import {BrowserRouter, Route, Redirect, Switch} from 'react-router-dom'
 import './App.css';
 
 /**
- * This file is temporary! It will be modified after routing and logic are set - Nikita
+ * File has been updated to use temporary router paths and display placeholder components - Nikita
  */
 
-const style = {
-  fill: '#61dafb',
-  fillOpacity: '0.125'
-}
-
-function App() {
-  let d9 = data.map(s => {
-    return {
-      x: parseFloat(s.current_salary.slice(1)),
-      y: parseFloat(s.previous_salary.slice(1)),
-    }
-  })
-  let dF = d9
-
-  dF.sort((a,b)=> a.x - b.x)
-
-  console.log("dF",dF)
-
-  let dP = {
-    label: 'somethingC',
-    values: [...dF]
-  }
-
-  let d0 = data.map(s => {
-    return {
-      x: parseFloat(s.current_salary.slice(1)),
-      y: parseFloat(s.previous_salary.slice(1)),
-    }
-  })
-
-  let dG = d0
-
-  dG.sort((a,b) => a.x - b.x)
-
-  console.log("dG",dG)
-
-  let dY = [].concat(...(data.map(s => {
-    return [
-      {
-        x: 'somethingA',
-        y: parseFloat(s.previous_salary.slice(1)),
-      },
-      {
-        x: 'somethingB',
-        y: parseFloat(s.current_salary.slice(1)),
-      }
-    ]
-  })))
-
-  let dX = {
-    label: 'somethingA',
-    values: [...dG]
-  }
-
-  let dZ = {
-    label: 'somethingA',
-    values: [...dY]
-  }
-
-  console.log(dX)
-  return (
-    <div >
-      <SideBar/>
-      <BarChart 
-        data={dZ} 
-        width={800}
-        height={600}
-        margin={{ top: 10, bottom: 50, left: 0, right: 10 }}
-      />
-      <hr />
-        <div className="sp">
-        <ScatterPlot
-                data={dX}
-                width={980}
-                height={600}
-                style={style}
-                margin={{top: 10, bottom: 50, left: 60, right: 10}}/>
-        </div>
-      <hr/>
-      <LineChart
-                data={dP}
-                width={1200}
-                height={600}
-                margin={{top: 10, bottom: 50, left: 60, right: 5}}/>
-    </div>
-  );
+const App = (props) => {
+    return (
+      <BrowserRouter>
+        <SideBar/>
+        <Switch>
+          <Redirect exact from="/" to="/datavisualization" component={Blank}/>
+          <Route path="/datavisualization" component={Blank}/>
+          <Route path="/charts" component={Charts}/>
+        </Switch>
+      </BrowserRouter>
+    )
 }
 
 export default App;
