+++ src/Blank.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Blank = (props) => {
+    return (
+      <div><h1>Blank</h1></div>
+    )
+  }
+
+export default Blank
